update docker compose and production image

+++ release.sh
@@ -1,19 +1,25 @@
-  #!/bin/bash
+#!/bin/bash
+BASE_DIR=$(dirname "$0")
+echo $BASE_DIR
+#cd /opt/efcsydney-roster/
+cd $BASE_DIR
+yarn
 
-  cd /opt/efcsydney-roster/
-  yarn
+pushd $BASE_DIR/client/
+yarn
+yarn build
+popd
 
-  pushd /opt/efcsydney-roster/client/
-  yarn
-  popd
-
-  echo "Clean up"
+echo "Clean up"
+if [ -f ./efcsydney-roster.tar.gz ]; then
   rm -rf efcsydney-roster.tar.gz
-  rm -rf client/node_modules
-  find log/* ! -name '.gitignore' -exec rm -rf {} +
+fi
 
-  echo "Tar"
-  pushd ../
-  tar -czf efcsydney-roster.tar.gz efcsydney-roster
-  mv efcsydney-roster.tar.gz efcsydney-roster/
-  popd
+rm -rf client/node_modules
+find log/* ! -name '.gitignore' -exec rm -rf {} +
+
+echo "Tar"
+pushd ../
+tar -czf efcsydney-roster.tar.gz efcsydney-roster
+mv efcsydney-roster.tar.gz efcsydney-roster/
+popd
